+++ src/main/java/com/sigmundgranaas/forgero/core/property/PropertyContainer.java
@@ -1,7 +1,5 @@
 package com.sigmundgranaas.forgero.core.property;
 
-import com.sigmundgranaas.forgero.core.property.attribute.Target;
-
 import java.util.List;
 
 public interface PropertyContainer {