+++ src/main/java/com/sigmundgranaas/forgero/core/property/active/ActivePropertyBuilder.java
@@ -6,7 +6,10 @@ import com.sigmundgranaas.forgero.core.property.PropertyPOJO;
 public class ActivePropertyBuilder {
     public static ActiveProperty createAttributeFromPojo(PropertyPOJO.Active activePOJO) {
         if (activePOJO.type == ActivePropertyType.BLOCK_BREAKING_PATTERN) {
-            return new PatternBreaking(activePOJO.pattern);
+            if (activePOJO.direction != null) {
+                return new PatternBreaking(activePOJO.pattern, activePOJO.direction);
+            }
+            return new PatternBreaking(activePOJO.pattern, BreakingDirection.ANY);
         } else {
             return new VeinBreaking(activePOJO.depth, activePOJO.tag, activePOJO.description);
         }