+++ src/main/java/com/sigmundgranaas/forgero/core/schematic/Schematic.java
@@ -2,7 +2,7 @@ package com.sigmundgranaas.forgero.core.schematic;
 
 import com.sigmundgranaas.forgero.core.property.Property;
 import com.sigmundgranaas.forgero.core.property.PropertyContainer;
-import com.sigmundgranaas.forgero.core.property.attribute.Target;
+import com.sigmundgranaas.forgero.core.property.Target;
 import com.sigmundgranaas.forgero.core.toolpart.ForgeroToolPartTypes;
 
 import java.util.List;