+++ src/main/java/com/sigmundgranaas/forgero/core/tool/ForgeroToolWithBinding.java
@@ -1,7 +1,7 @@
 package com.sigmundgranaas.forgero.core.tool;
 
 import com.sigmundgranaas.forgero.core.property.Property;
-import com.sigmundgranaas.forgero.core.property.attribute.Target;
+import com.sigmundgranaas.forgero.core.property.Target;
 import com.sigmundgranaas.forgero.core.toolpart.binding.ToolPartBinding;
 import com.sigmundgranaas.forgero.core.toolpart.handle.ToolPartHandle;
 import com.sigmundgranaas.forgero.core.toolpart.head.ToolPartHead;