+++ src/main/java/com/sigmundgranaas/forgero/item/items/tool/ForgeroAxeItem.java
@@ -1,7 +1,7 @@
 package com.sigmundgranaas.forgero.item.items.tool;
 
 import com.sigmundgranaas.forgero.ForgeroInitializer;
-import com.sigmundgranaas.forgero.core.property.attribute.Target;
+import com.sigmundgranaas.forgero.core.property.Target;
 import com.sigmundgranaas.forgero.core.tool.ForgeroTool;
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolTypes;
 import com.sigmundgranaas.forgero.core.toolpart.handle.ToolPartHandle;
@@ -48,6 +48,16 @@ public class ForgeroAxeItem extends AxeItem implements ForgeroToolItem {
     }
 
     @Override
+    public Text getName() {
+        return getForgeroTranslatableToolName();
+    }
+
+    @Override
+    public Text getName(ItemStack stack) {
+        return getForgeroTranslatableToolName();
+    }
+
+    @Override
     public ForgeroToolTypes getToolType() {
         return ForgeroToolTypes.AXE;
     }
@@ -81,6 +91,6 @@ public class ForgeroAxeItem extends AxeItem implements ForgeroToolItem {
     protected String getOrCreateTranslationKey() {
         return String.format("item.%s.%s_%s", ForgeroInitializer.MOD_NAMESPACE, tool.getToolHead().getPrimaryMaterial().getName(), getToolType().toString().toLowerCase(Locale.ROOT));
     }
-    
+
 
 }