+++ src/main/java/com/sigmundgranaas/forgero/item/items/tool/ForgeroHoeItem.java
@@ -1,7 +1,7 @@
 package com.sigmundgranaas.forgero.item.items.tool;
 
 import com.sigmundgranaas.forgero.ForgeroInitializer;
-import com.sigmundgranaas.forgero.core.property.attribute.Target;
+import com.sigmundgranaas.forgero.core.property.Target;
 import com.sigmundgranaas.forgero.core.tool.ForgeroTool;
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolTypes;
 import com.sigmundgranaas.forgero.core.toolpart.handle.ToolPartHandle;
@@ -11,6 +11,7 @@ import com.sigmundgranaas.forgero.item.adapter.DescriptionWriter;
 import com.sigmundgranaas.forgero.item.adapter.FabricToForgeroToolAdapter;
 import net.minecraft.block.Block;
 import net.minecraft.client.item.TooltipContext;
+import net.minecraft.item.HoeItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
 import net.minecraft.tag.BlockTags;
@@ -23,13 +24,13 @@ import java.util.List;
 import java.util.Locale;
 
 
-public class ShovelItem extends net.minecraft.item.ShovelItem implements ForgeroToolItem {
+public class ForgeroHoeItem extends HoeItem implements ForgeroToolItem {
 
     private final ForgeroTool tool;
     private final FabricToForgeroToolAdapter toolAdapter = FabricToForgeroToolAdapter.createAdapter();
 
 
-    public ShovelItem(ToolMaterial toolMaterial, Settings settings, ForgeroTool tool) {
+    public ForgeroHoeItem(ToolMaterial toolMaterial, Settings settings, ForgeroTool tool) {
         super(toolMaterial, (int) tool.getAttackDamage(Target.createEmptyTarget()), tool.getAttackSpeed(Target.createEmptyTarget()), settings);
         this.tool = tool;
     }
@@ -41,15 +42,26 @@ public class ShovelItem extends net.minecraft.item.ShovelItem implements Forgero
         forgeroTool.createToolDescription(new DescriptionWriter(tooltip));
         super.appendTooltip(itemStack, world, tooltip, tooltipContext);
     }
-    
+
     @Override
     public ForgeroToolTypes getToolType() {
-        return ForgeroToolTypes.SHOVEL;
+        return ForgeroToolTypes.HOE;
     }
 
     @Override
     public ForgeroTool getTool() {
         return tool;
+    }
+
+
+    @Override
+    public Text getName() {
+        return getForgeroTranslatableToolName();
+    }
+
+    @Override
+    public Text getName(ItemStack stack) {
+        return getForgeroTranslatableToolName();
     }
 
 
@@ -75,7 +87,7 @@ public class ShovelItem extends net.minecraft.item.ShovelItem implements Forgero
 
     @Override
     public TagKey<Block> getToolTags() {
-        return BlockTags.SHOVEL_MINEABLE;
+        return BlockTags.HOE_MINEABLE;
     }
 
     @Override