+++ src/main/java/com/sigmundgranaas/forgero/registry/ItemRegistryImpl.java
@@ -33,6 +33,7 @@ public class ItemRegistryImpl implements ItemRegistry {
 
     public static final TagKey<Block> VEIN_MINING_ORES = TagKey.of(Registry.BLOCK_KEY, new Identifier(ForgeroInitializer.MOD_NAMESPACE, "vein_mining_ores"));
     public static final TagKey<Block> VEIN_MINING_SAND = TagKey.of(Registry.BLOCK_KEY, new Identifier(ForgeroInitializer.MOD_NAMESPACE, "vein_mining_sand"));
+    public static final TagKey<Block> PLANTS = TagKey.of(Registry.BLOCK_KEY, new Identifier(ForgeroInitializer.MOD_NAMESPACE, "plants"));
 
 
     private static ItemRegistry INSTANCE;